pyproject.toml
[tool.poetry]
name = "another"
version = "0.1.0"
description = ""
authors = ["Dominik Kozaczko <dominik@kozaczko.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
django-getpaid = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
